--- server/database/seeders/20211201125158-demo-menu.js
@@ -1,1690 +1,0 @@
-let i = 69;
-function autoIncrement() {
-  i = i + 1;
-}
-
-module.exports = {
-  up: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkInsert(
-      'menu',
-      [
-        {
-          id: 1,
-          menu_name: '눈꽃치즈 떡볶이',
-          menu_price: 15000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%82%E1%85%AE%E1%86%AB%E1%84%81%E1%85%A9%E1%86%BE%E1%84%8E%E1%85%B5%E1%84%8C%E1%85%B3%E1%84%84%E1%85%A5%E1%86%A8%E1%84%87%E1%85%A9%E1%86%A9%E1%84%8B%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 1,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 2,
-          menu_name: '카레 떡볶이',
-          menu_price: 11000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8F%E1%85%A1%E1%84%85%E1%85%A6%E1%84%84%E1%85%A5%E1%86%A8%E1%84%87%E1%85%A9%E1%86%A9%E1%84%8B%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 1,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 3,
-          menu_name: '크림떡볶이',
-          menu_price: 13000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/tteokbokki.jpeg',
-          menu_order_quantity: 0,
-          store_id: 1,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 4,
-          menu_name: '빨계떡라면',
-          menu_price: 4500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%85%E1%85%A1%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 2,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 5,
-          menu_name: '계떡 라면',
-          menu_price: 4500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%85%E1%85%A1%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 2,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 6,
-          menu_name: '치즈 빨계부대 라면',
-          menu_price: 6500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%85%E1%85%A1%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 2,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 7,
-          menu_name: '일품 멸치국수',
-          menu_price: 5500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A9%E1%86%AF%E1%84%87%E1%85%A2%E1%86%BC%E1%84%8B%E1%85%B5%E1%84%87%E1%85%B5%E1%84%87%E1%85%B5%E1%86%B7%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 3,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 8,
-          menu_name: '매콤새콤 비빔국수',
-          menu_price: 5500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A9%E1%86%AF%E1%84%87%E1%85%A2%E1%86%BC%E1%84%8B%E1%85%B5%E1%84%87%E1%85%B5%E1%84%87%E1%85%B5%E1%86%B7%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 3,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 9,
-          menu_name: '달콤고소 검은콩국수',
-          menu_price: 15000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A9%E1%86%AF%E1%84%87%E1%85%A2%E1%86%BC%E1%84%8B%E1%85%B5%E1%84%87%E1%85%B5%E1%84%87%E1%85%B5%E1%86%B7%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 3,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 10,
-          menu_name: '매콤 국물 떡볶이',
-          menu_price: 3500,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/tteokbokki.jpeg',
-          menu_order_quantity: 0,
-          store_id: 4,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 11,
-          menu_name: '긴다리 오징어튀김(1개)',
-          menu_price: 1900,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%A9%E1%84%8C%E1%85%B5%E1%86%BC%E1%84%8B%E1%85%A5%E1%84%90%E1%85%B1%E1%84%80%E1%85%B5%E1%86%B7.jpg',
-          menu_order_quantity: 0,
-          store_id: 4,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 12,
-          menu_name: '순대(1인분)',
-          menu_price: 2500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%AE%E1%86%AB%E1%84%83%E1%85%A2.jpg',
-          menu_order_quantity: 0,
-          store_id: 4,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 13,
-          menu_name: '웅이네 제육덮밥',
-          menu_price: 6000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%A6%E1%84%8B%E1%85%B2%E1%86%A8%E1%84%87%E1%85%A9%E1%86%A9%E1%84%8B%E1%85%B3%E1%86%B7.jpg',
-          menu_order_quantity: 0,
-          store_id: 5,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 14,
-          menu_name: '웅이네 김치볶음밥',
-          menu_price: 5500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%A6%E1%84%8B%E1%85%B2%E1%86%A8%E1%84%87%E1%85%A9%E1%86%A9%E1%84%8B%E1%85%B3%E1%86%B7.jpg',
-          menu_order_quantity: 0,
-          store_id: 5,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 15,
-          menu_name: '웅이네 비빔밥',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%87%E1%85%B5%E1%84%87%E1%85%B5%E1%86%B7%E1%84%87%E1%85%A1%E1%86%B8.jpg',
-          menu_order_quantity: 0,
-          store_id: 5,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 16,
-          menu_name: '웅이 김밥',
-          menu_price: 2200,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%B5%E1%86%B7%E1%84%87%E1%85%A1%E1%86%B8.jpg',
-          menu_order_quantity: 0,
-          store_id: 5,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 17,
-          menu_name: '들깨 칼국수',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8F%E1%85%A1%E1%86%AF%E1%84%80%E1%85%AE%E1%86%A8%E1%84%89%E1%85%AE.jpg',
-          menu_order_quantity: 0,
-          store_id: 6,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 18,
-          menu_name: '칼 수제비',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8F%E1%85%A1%E1%86%AF%E1%84%80%E1%85%AE%E1%86%A8%E1%84%89%E1%85%AE.jpg',
-          menu_order_quantity: 0,
-          store_id: 6,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 19,
-          menu_name: '해물 칼국수',
-          menu_price: 9000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8F%E1%85%A1%E1%86%AF%E1%84%80%E1%85%AE%E1%86%A8%E1%84%89%E1%85%AE.jpg',
-          menu_order_quantity: 0,
-          store_id: 6,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 20,
-          menu_name: '마르게리타 피자',
-          menu_price: 11500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/pizza/pizza66.jpg',
-          menu_order_quantity: 0,
-          store_id: 7,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 21,
-          menu_name: '루꼴라 피자',
-          menu_price: 13500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%90%E1%85%A9%E1%84%86%E1%85%A1%E1%84%90%E1%85%A9%E1%84%91%E1%85%B5%E1%84%8C%E1%85%A1.jpg',
-          menu_order_quantity: 0,
-          store_id: 7,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 22,
-          menu_name: '토마토 치즈 스파게티',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%87%E1%85%A9%E1%86%AF%E1%84%85%E1%85%A9%E1%84%82%E1%85%A6%E1%84%8C%E1%85%A6%E1%84%91%E1%85%A1%E1%84%89%E1%85%B3%E1%84%90%E1%85%A1.jpg',
-          menu_order_quantity: 0,
-          store_id: 7,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 23,
-          menu_name: '아보카도계란 샐러드',
-          menu_price: 400,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%A1%E1%84%87%E1%85%A9%E1%84%8F%E1%85%A1%E1%84%83%E1%85%A9%E1%84%89%E1%85%A2%E1%86%AF%E1%84%85%E1%85%A5%E1%84%83%E1%85%B3.jpg',
-          menu_order_quantity: 0,
-          store_id: 8,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 24,
-          menu_name: '연어 샐러드',
-          menu_price: 11000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%A1%E1%84%87%E1%85%A9%E1%84%8F%E1%85%A1%E1%84%83%E1%85%A9%E1%84%89%E1%85%A2%E1%86%AF%E1%84%85%E1%85%A5%E1%84%83%E1%85%B3.jpg',
-          menu_order_quantity: 0,
-          store_id: 8,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 25,
-          menu_name: '리코타 치즈 샐러드',
-          menu_price: 9000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%91%E1%85%A1%E1%84%89%E1%85%B3%E1%84%90%E1%85%A1%E1%84%89%E1%85%A2%E1%86%AF%E1%84%85%E1%85%A5%E1%84%83%E1%85%B3.jpg',
-          menu_order_quantity: 0,
-          store_id: 8,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        // {
-        //   id: 26,
-        //   menu_name: '알리오올리오 파스타',
-        //   menu_price: 11000,
-        //   menu_image: 'https://foodish-api.herokuapp.com/images/pasta/pasta28.jpg',
-        //   menu_order_quantity: 0,
-        //   store_id: 9,
-        //   created_at: new Date(),
-        //   updated_at: new Date(),
-        // },
-        // {
-        //   id: 27,
-        //   menu_name: '까르보나라 파스타',
-        //   menu_price: 11000,
-        //   menu_image: 'https://foodish-api.herokuapp.com/images/pasta/pasta28.jpg',
-        //   menu_order_quantity: 0,
-        //   store_id: 9,
-        //   created_at: new Date(),
-        //   updated_at: new Date(),
-        // },
-        // {
-        //   id: 28,
-        //   menu_name: '링귀니 볼로네제 파스타',
-        //   menu_price: 12000,
-        //   menu_image: 'https://foodish-api.herokuapp.com/images/pasta/pasta28.jpg',
-        //   menu_order_quantity: 0,
-        //   store_id: 9,
-        //   created_at: new Date(),
-        //   updated_at: new Date(),
-        // },
-        {
-          id: 29,
-          menu_name: '양고기 스튜',
-          menu_price: 15000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/lamb-g446b05bae_640.jpg',
-          menu_order_quantity: 0,
-          store_id: 10,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 30,
-          menu_name: '지중해식 양고기 바베큐',
-          menu_price: 19500,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/food-g92e71fad0_640.jpg',
-          menu_order_quantity: 0,
-          store_id: 10,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 31,
-          menu_name: '양고기 치아바타',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/sandwich-g80e4f711a_640.jpg',
-          menu_order_quantity: 0,
-          store_id: 10,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 32,
-          menu_name: '트러플 크림파스타와 고르곤졸라 피자(세트A)',
-          menu_price: 32000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store3.png',
-          menu_order_quantity: 0,
-          store_id: 11,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 33,
-          menu_name: '게살마늘 크림파스타와 디아볼라 피자(세트B)',
-          menu_price: 33000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store3.png',
-          menu_order_quantity: 0,
-          store_id: 11,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 34,
-          menu_name: '페스카 크레마파스타와 핫 쉬림프 피자(세트C)',
-          menu_price: 36500,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store3.png',
-          menu_order_quantity: 0,
-          store_id: 11,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 35,
-          menu_name: '리스토란테 A코스',
-          menu_price: 107000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store1.png',
-          menu_order_quantity: 0,
-          store_id: 12,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 36,
-          menu_name: '리스토란테 B코스',
-          menu_price: 111000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store1.png',
-          menu_order_quantity: 0,
-          store_id: 12,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 37,
-          menu_name: '불족 보쌈(대)',
-          menu_price: 32000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%86%AB%E1%84%83%E1%85%AE%E1%84%8C%E1%85%A5%E1%86%AB%E1%84%80%E1%85%A9%E1%86%AF.jpg',
-          menu_order_quantity: 0,
-          store_id: 13,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 38,
-          menu_name: '마늘 보쌈(대)',
-          menu_price: 32000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%86%AB%E1%84%83%E1%85%AE%E1%84%8C%E1%85%A5%E1%86%AB%E1%84%80%E1%85%A9%E1%86%AF.jpg',
-          menu_order_quantity: 0,
-          store_id: 13,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 39,
-          menu_name: '왕새우튀김 + 황도',
-          menu_price: 17000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store2.png',
-          menu_order_quantity: 0,
-          store_id: 14,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 40,
-          menu_name: '두부김치제육 + 황도',
-          menu_price: 17500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%A6%E1%84%8B%E1%85%B2%E1%86%A8%E1%84%87%E1%85%A9%E1%86%A9%E1%84%8B%E1%85%B3%E1%86%B7.jpg',
-          menu_order_quantity: 0,
-          store_id: 14,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 41,
-          menu_name: '계란말이 + 김치전 + 황도',
-          menu_price: 20000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%B5%E1%86%B7%E1%84%8E%E1%85%B5%E1%84%8C%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 14,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 42,
-          menu_name: '닭도리탕',
-          menu_price: 25000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A1%E1%86%B7%E1%84%8C%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 15,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 43,
-          menu_name: '알탕',
-          menu_price: 18000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A1%E1%86%B7%E1%84%8C%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 15,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 44,
-          menu_name: '골뱅이소면무침',
-          menu_price: 16000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A9%E1%86%AF%E1%84%87%E1%85%A2%E1%86%BC%E1%84%8B%E1%85%B5%E1%84%87%E1%85%B5%E1%84%87%E1%85%B5%E1%86%B7%E1%84%86%E1%85%A7%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 15,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 45,
-          menu_name: '옛날 통닭 + 생맥주(1L)',
-          menu_price: 27000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8E%E1%85%B5%E1%84%8F%E1%85%B5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 16,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 46,
-          menu_name: '먹태 + 생맥주(1L)',
-          menu_price: 18000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8E%E1%85%B5%E1%84%8F%E1%85%B5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 16,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 47,
-          menu_name: '모듬꼬치(소,양,닭) + 생맥주(1L)',
-          menu_price: 22000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8E%E1%85%B5%E1%84%8F%E1%85%B5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 16,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 48,
-          menu_name: '후라이드 치킨',
-          menu_price: 16000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8E%E1%85%B5%E1%84%8F%E1%85%B5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 17,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 49,
-          menu_name: '간장 치킨',
-          menu_price: 18000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/butter-chicken/butter-chicken10.jpg',
-          menu_order_quantity: 0,
-          store_id: 17,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 50,
-          menu_name: '마늘 치킨',
-          menu_price: 18000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/butter-chicken/butter-chicken10.jpg',
-          menu_order_quantity: 0,
-          store_id: 17,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 51,
-          menu_name: '연탄 돼지갈비(600g)',
-          menu_price: 26000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/butter-chicken/butter-chicken10.jpg',
-          menu_order_quantity: 0,
-          store_id: 18,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 52,
-          menu_name: '특급 삼겹살(600g)',
-          menu_price: 26000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%A1%E1%86%B7%E1%84%80%E1%85%A7%E1%86%B8%E1%84%89%E1%85%A1%E1%86%AF.jpg',
-          menu_order_quantity: 0,
-          store_id: 18,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 53,
-          menu_name: '두루치기(600g)',
-          menu_price: 28000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%A1%E1%86%B7%E1%84%80%E1%85%A7%E1%86%B8%E1%84%89%E1%85%A1%E1%86%AF.jpg',
-          menu_order_quantity: 0,
-          store_id: 18,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 54,
-          menu_name: '생크림 단팥빵',
-          menu_price: 1600,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 19,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 55,
-          menu_name: '녹차 크림 단팥빵',
-          menu_price: 2000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/dessert/dessert18.jpg',
-          menu_order_quantity: 0,
-          store_id: 19,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 56,
-          menu_name: '아메리카노 + 딸기 쇼트 케이크',
-          menu_price: 5800,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8F%E1%85%A5%E1%86%B8%E1%84%8F%E1%85%A6%E1%84%8B%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 20,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 57,
-          menu_name: '아메리카노 + 호두 파운드 케이크',
-          menu_price: 6200,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8F%E1%85%A5%E1%86%B8%E1%84%8F%E1%85%A6%E1%84%8B%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 20,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 58,
-          menu_name: '민트 초코 스무디',
-          menu_price: 3200,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/kiwi-coctail-ge03797d53_640.jpg',
-          menu_order_quantity: 0,
-          store_id: 21,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 59,
-          menu_name: '딸기 바나나 스무디',
-          menu_price: 3000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/dessert/dessert18.jpg',
-          menu_order_quantity: 0,
-          store_id: 21,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 60,
-          menu_name: '블루베리 요거트 스무디',
-          menu_price: 3500,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%86%E1%85%AE%E1%84%83%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 21,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 61,
-          menu_name: '찹쌀 모짜렐라 콘도그(1개)',
-          menu_price: 2000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 22,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 62,
-          menu_name: '찹쌀꽈배기(4개)',
-          menu_price: 3000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 22,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 63,
-          menu_name: '찹쌀 자몽도너츠(1개)',
-          menu_price: 1100,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 22,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 64,
-          menu_name: '초코분수 식빵',
-          menu_price: 2900,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B5%E1%86%A8%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 23,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 65,
-          menu_name: '블루베리 크림치즈식빵',
-          menu_price: 2900,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B5%E1%86%A8%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 23,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 66,
-          menu_name: '호두 밤 식빵',
-          menu_price: 2900,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B5%E1%86%A8%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 23,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 67,
-          menu_name: '사라다 빵',
-          menu_price: 1300,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 24,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 68,
-          menu_name: '소보루 밤',
-          menu_price: 1100,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 24,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: 69,
-          menu_name: '완두 앙금 빵',
-          menu_price: 1600,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%88%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 24,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '스키다시',
-          menu_price: 4000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 25,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '오코노미야끼',
-          menu_price: 10000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%B5%E1%86%AF%E1%84%89%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 25,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '규동',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%B5%E1%86%AF%E1%84%89%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 25,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '오늘의 초밥세트',
-          menu_price: 13000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 25,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '붓가케우동',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%AE%E1%84%83%E1%85%A9%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 25,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '모듬초밥',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 26,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '스파이시튜나롤',
-          menu_price: 11000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 26,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '와규초밥',
-          menu_price: 14000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 26,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '와규 타다키',
-          menu_price: 20000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%B5%E1%86%AF%E1%84%89%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 27,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '도리 가라아게',
-          menu_price: 12000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%B5%E1%86%AF%E1%84%89%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 27,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '메로 구이',
-          menu_price: 21000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%B5%E1%86%AF%E1%84%89%E1%85%B5%E1%86%A8.jpg',
-          menu_order_quantity: 0,
-          store_id: 27,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '모츠나베',
-          menu_price: 20000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%AE%E1%84%83%E1%85%A9%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 27,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '어부지리',
-          menu_price: 20000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%A2%E1%86%BC%E1%84%89%E1%85%A5%E1%86%AB%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 28,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '복지리',
-          menu_price: 23000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%A2%E1%86%BC%E1%84%89%E1%85%A5%E1%86%AB%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 28,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '런치스페셜',
-          menu_price: 25000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store5.png',
-          menu_order_quantity: 0,
-          store_id: 28,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '새송이구이',
-          menu_price: 30000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%87%E1%85%A5%E1%84%89%E1%85%A5%E1%86%BA.jpg',
-          menu_order_quantity: 0,
-          store_id: 28,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '참치회',
-          menu_price: 17000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%92%E1%85%AC.jpg',
-          menu_order_quantity: 0,
-          store_id: 28,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '참치뱃살',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%92%E1%85%AC.jpg',
-          menu_order_quantity: 0,
-          store_id: 29,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '도미회',
-          menu_price: 12000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%92%E1%85%AC.jpg',
-          menu_order_quantity: 0,
-          store_id: 29,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '고등어회',
-          menu_price: 13000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%92%E1%85%AC.jpg',
-          menu_order_quantity: 0,
-          store_id: 29,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '숭어회',
-          menu_price: 14000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%92%E1%85%AC.jpg',
-          menu_order_quantity: 0,
-          store_id: 29,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '런치초밥세트(6p)',
-          menu_price: 12000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 30,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '초밥세트(10p)',
-          menu_price: 12000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 30,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '특초밥세트(12p)',
-          menu_price: 13000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 30,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '스페셜초밥세트(13p)',
-          menu_price: 13000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%B3%E1%84%89%E1%85%B5.jpg',
-          menu_order_quantity: 0,
-          store_id: 30,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라탕',
-          menu_price: 10000,
-          menu_order_quantity: 0,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          store_id: 31,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '지삼선',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 31,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '꿔바로우',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store4.png',
-          menu_order_quantity: 0,
-          store_id: 31,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '미씨엔',
-          menu_price: 10000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8B%E1%85%AE%E1%84%83%E1%85%A9%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 32,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '셩젠바오',
-          menu_price: 11000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%89%E1%85%A7%E1%86%BC%E1%84%8C%E1%85%A6%E1%86%AB%E1%84%87%E1%85%A1%E1%84%8B%E1%85%A9.jpg',
-          menu_order_quantity: 0,
-          store_id: 32,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '어향가지',
-          menu_price: 10000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 32,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '샨시량피',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 32,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라샹궈',
-          menu_price: 12000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 32,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라탕',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC2.jpg',
-          menu_order_quantity: 0,
-          store_id: 33,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '지삼선',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC2.jpg',
-          menu_order_quantity: 0,
-          store_id: 33,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '꿔바로우',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store4.png',
-          menu_order_quantity: 0,
-          store_id: 33,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라탕',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC2.jpg',
-          menu_order_quantity: 0,
-          store_id: 34,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '지삼선',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC2.jpg',
-          menu_order_quantity: 0,
-          store_id: 34,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '꿔바로우',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store4.png',
-          menu_order_quantity: 0,
-          store_id: 34,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라탕',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC2.jpg',
-          menu_order_quantity: 0,
-          store_id: 35,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '지삼선',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 35,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '꿔바로우',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store4.png',
-          menu_order_quantity: 0,
-          store_id: 35,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라탕',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%86%E1%85%A1%E1%84%85%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC2.jpg',
-          menu_order_quantity: 0,
-          store_id: 36,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '지삼선',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 36,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '꿔바로우',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store4.png',
-          menu_order_quantity: 0,
-          store_id: 36,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마라탕',
-          menu_price: 7000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 36,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '지삼선',
-          menu_price: 8000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%8C%E1%85%B5%E1%84%89%E1%85%A1%E1%86%B7%E1%84%89%E1%85%A5%E1%86%AB.jpg',
-          menu_order_quantity: 0,
-          store_id: 36,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '꿔바로우',
-          menu_price: 10000,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/store4.png',
-          menu_order_quantity: 0,
-          store_id: 36,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '징거버거',
-          menu_price: 7000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 37,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '다빈치버거',
-          menu_price: 6500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger80.jpg',
-          menu_order_quantity: 0,
-          store_id: 37,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마마버거',
-          menu_price: 6800,
-          menu_image: 'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/burger-g1ccf49c3e_640.jpg',
-          menu_order_quantity: 0,
-          store_id: 37,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '징거버거',
-          menu_price: 7000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger80.jpg',
-          menu_order_quantity: 0,
-          store_id: 38,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '다빈치버거',
-          menu_price: 6500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger80.jpg',
-          menu_order_quantity: 0,
-          store_id: 38,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마마버거',
-          menu_price: 6800,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger80.jpg',
-          menu_order_quantity: 0,
-          store_id: 38,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '징거버거',
-          menu_price: 7000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger80.jpg',
-          menu_order_quantity: 0,
-          store_id: 39,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '다빈치버거',
-          menu_price: 6500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger80.jpg',
-          menu_order_quantity: 0,
-          store_id: 39,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마마버거',
-          menu_price: 6800,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 39,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '징거버거',
-          menu_price: 7000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 40,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '다빈치버거',
-          menu_price: 6500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 40,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마마버거',
-          menu_price: 6800,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 40,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '징거버거',
-          menu_price: 7000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 41,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '다빈치버거',
-          menu_price: 6500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 41,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마마버거',
-          menu_price: 6800,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 41,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '징거버거',
-          menu_price: 7000,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 42,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '다빈치버거',
-          menu_price: 6500,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 42,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '마마버거',
-          menu_price: 6800,
-          menu_image: 'https://foodish-api.herokuapp.com/images/burger/burger87.jpg',
-          menu_order_quantity: 0,
-          store_id: 42,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%83%E1%85%AC%E1%86%AB%E1%84%8C%E1%85%A1%E1%86%BC%E1%84%8D%E1%85%B5%E1%84%80%E1%85%A2.jpg',
-          menu_order_quantity: 0,
-          store_id: 43,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A1%E1%86%B7%E1%84%8C%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 43,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%80%E1%85%A1%E1%86%B7%E1%84%8C%E1%85%A1%E1%84%90%E1%85%A1%E1%86%BC.jpg',
-          menu_order_quantity: 0,
-          store_id: 43,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image:
-            'https://meal2sdk.s3.ap-northeast-2.amazonaws.com/dummyimage/%E1%84%83%E1%85%AC%E1%86%AB%E1%84%8C%E1%85%A1%E1%86%BC%E1%84%8D%E1%85%B5%E1%84%80%E1%85%A2.jpg',
-          menu_order_quantity: 0,
-          store_id: 44,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 44,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 44,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 45,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 45,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 45,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 46,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 46,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 46,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 47,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 47,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 47,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 48,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 48,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 48,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 49,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 49,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 49,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '된장찌개',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 50,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '김치찌개',
-          menu_price: 5000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 50,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-        {
-          id: autoIncrement(),
-          menu_name: '육개장',
-          menu_price: 6000,
-          menu_image: 'https://meal2sdk.s3.amazonaws.com/-001_12.jpg',
-          menu_order_quantity: 0,
-          store_id: 50,
-          created_at: new Date(),
-          updated_at: new Date(),
-        },
-      ],
-      {}
-    );
-  },
-
-  down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete('menu', null, {});
-  },
-};
